Multiple changes to the hub and library involving shutdown coordination and variable handling.
More information is now returned with variable requests to the hub, and the library uses this information to perform a relevant action. Variable errors are handled on the client side and read only variable reads are cached. The library will also handle being asychronously shutdown by the hub in the event of an error requiring the hub to drop a client.